1. Updated naming and namespace 2. Static events now remove all listeners on destroy
