+++ Assets/Microgames/DatingSim/Scripts/DatingSimDialoguePreset.cs.meta
