+++ Assets/Microgames/DatingSim/Scripts/DatingSimGameController.cs
@@ -2,16 +2,18 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class DatingSimGameController : MonoBehaviour {
+namespace NitorInc.DatingSim {
+    public class DatingSimGameController : MonoBehaviour {
 
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
+        // Use this for initialization
+        void Start() {
 
+        }
+
+        // Update is called once per frame
+        void Update() {
+
+        }
+
+    }
 }
